vhd: replace atomicio with in-tree io-util helpers
The journal used atomicio() from atomicio.c/.h, which carry a separate
OpenBSD license. Replace it with vhd_atomic_io() / vhd_atomic_pio() in
the new io-util.c, under the same license as the rest of the tree.

vhd_atomic_pio() is the positioned helper moved out of libvhd.c;
vhd_atomic_io() is the sequential read/write variant.
